dynamicModels/requirements.txt
asgiref==3.4.1
Django==3.0.3
django-extensions==3.2.1
djangorestframework==3.14.0
pkg-resources==0.0.0
psycopg2==2.8.6
pytz==2023.2
sqlparse==0.4.3
typing-extensions==4.1.1
